+++ source/config/Options.ts
@@ -209,7 +209,7 @@ export class Options {
 
     {
       brand: OptionBrand.String,
-      description: "The look up strategy to be used to find the TSConfig file.",
+      description: "The TSConfig to load.",
       group: OptionGroup.CommandLine | OptionGroup.ConfigFile | OptionGroup.Task,
       name: "tsconfig",
     },
@@ -264,7 +264,11 @@ export class Options {
   }
 
   static #isLookupStrategy(optionValue: string) {
-    return ["findup", "ignore"].includes(optionValue);
+    return ["findup", "baseline"].includes(optionValue);
+  }
+
+  static isJsonString(text: string): boolean {
+    return text.startsWith("{");
   }
 
   static resolve(optionName: string, optionValue: string, rootPath = "."): string {
@@ -274,7 +278,10 @@ export class Options {
       case "config":
       case "root":
       case "tsconfig":
-        if (canonicalOptionName === "tsconfig" && Options.#isLookupStrategy(optionValue)) {
+        if (
+          canonicalOptionName === "tsconfig" &&
+          (Options.#isLookupStrategy(optionValue) || Options.isJsonString(optionValue))
+        ) {
           break;
         }
 
@@ -319,7 +326,10 @@ export class Options {
       case "config":
       case "root":
       case "tsconfig":
-        if (canonicalOptionName === "tsconfig" && Options.#isLookupStrategy(optionValue)) {
+        if (
+          canonicalOptionName === "tsconfig" &&
+          (Options.#isLookupStrategy(optionValue) || Options.isJsonString(optionValue))
+        ) {
           break;
         }