+++ source/result/ProjectConfigKind.enum.ts
@@ -1,0 +1,6 @@
+export const enum ProjectConfigKind {
+  Discovered,
+  Default,
+  Provided,
+  Synthetic,
+}